requirements.txt
altgraph==0.17.4
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
frida==17.2.11
frida-tools==14.4.1
idna==3.10
MouseInfo==0.1.3
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
prompt_toolkit==3.0.51
psutil==7.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
urllib3==2.5.0
wcwidth==0.2.13
websockets==13.1
